+++ app/src/main/java/eu/kanade/tachiyomi/util/system/DeviceUtil.kt
@@ -1,15 +1,17 @@
 package eu.kanade.tachiyomi.util.system
 
 import android.annotation.SuppressLint
-import android.content.Context
+import android.app.Activity
 import android.hardware.display.DisplayManager
 import android.os.Build
 import android.view.Display
-import android.view.View
 import android.view.Window
-import android.view.WindowManager
 import androidx.core.content.getSystemService
+import androidx.core.view.WindowInsetsCompat
 import timber.log.Timber
+import java.lang.reflect.InvocationTargetException
+
+
 
 object DeviceUtil {
 
@@ -92,59 +94,104 @@ object DeviceUtil {
     fun setLegacyCutoutMode(window: Window, mode: LegacyCutoutMode) {
         when (mode) {
             LegacyCutoutMode.SHORT_EDGES -> {
-                /* Deprecated method
+                // Vivo doesn't support this, user had to set it from Settings
+                /*
                 if (isVivo) {
-                    window.addFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN)
-                    window.addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
-                    var systemUiVisibility = window.decorView.systemUiVisibility
-                    systemUiVisibility = systemUiVisibility or View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                    systemUiVisibility = systemUiVisibility or View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                    window.decorView.systemUiVisibility = systemUiVisibility
                 }
                  */
             }
             LegacyCutoutMode.NEVER -> {
-                /* Deprecated method
+                // Vivo doesn't support this, user had to set it from Settings
+                /*
                 if (isVivo) {
-                    window.clearFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN)
-                    window.clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
-                    var systemUiVisibility = window.decorView.systemUiVisibility
-                    systemUiVisibility = systemUiVisibility and View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN.inv()
-                    systemUiVisibility = systemUiVisibility and View.SYSTEM_UI_FLAG_LAYOUT_STABLE.inv()
-                    window.decorView.systemUiVisibility = systemUiVisibility
                 }
                  */
             }
         }
     }
 
-    fun hasCutout(context: Context? = null): Boolean {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-                return context?.getSystemService<DisplayManager>()
-                    ?.getDisplay(Display.DEFAULT_DISPLAY)?.cutout != null
-            }
-            // TODO: Actually check for cutout
-            return true
-        }
-        /*
-        else if (isVivo && context != null) {
+    fun hasCutout(context: Activity?): CutoutSupport {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            if (context?.getSystemService<DisplayManager>()
+                    ?.getDisplay(Display.DEFAULT_DISPLAY)?.cutout != null)
+                return CutoutSupport.EXTENDED
+        } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+            val displayCutout = context?.window?.decorView?.rootWindowInsets?.displayCutout
+            if (displayCutout?.safeInsetTop != null || displayCutout?.safeInsetBottom != null)
+                return CutoutSupport.MODERN
+        } else if (isVivo) {
             // https://swsdl.vivo.com.cn/appstore/developer/uploadfile/20180328/20180328152252602.pdf
             try {
                 @SuppressLint("PrivateApi")
-                val ftFeature = context.classLoader
-                    .loadClass("android.util.FtFeature")
-                val isFeatureSupportMethod = ftFeature.getMethod(
+                val ftFeature = context?.classLoader
+                    ?.loadClass("android.util.FtFeature")
+                val isFeatureSupportMethod = ftFeature?.getMethod(
                     "isFeatureSupport",
-                    Int::class.javaPrimitiveType
+                    Int::class.javaPrimitiveType,
                 )
                 val isNotchOnScreen = 0x00000020
-                return isFeatureSupportMethod.invoke(ftFeature, isNotchOnScreen) as Boolean
+                val isSupported = isFeatureSupportMethod?.invoke(ftFeature, isNotchOnScreen) as Boolean
+                if (isSupported) return CutoutSupport.LEGACY
+            } catch (_: Exception) {
+            }
+        } else if (isMiui) {
+            try {
+                @SuppressLint("PrivateApi")
+                val sysProp = context?.classLoader?.loadClass("android.os.SystemProperties")
+                val method = sysProp?.getMethod("getInt", String::class.java, Int::class.javaPrimitiveType)
+                val rt = method?.invoke(sysProp, "ro.miui.notch", 0) as Int
+                if (rt == 1) return CutoutSupport.LEGACY
             } catch (_: Exception) {
             }
         }
-        */
-        return false
+        return CutoutSupport.NONE
+    }
+
+    fun getCutoutHeight(context: Activity?, cutoutSupport: CutoutSupport): Number {
+        return when (cutoutSupport) {
+            CutoutSupport.MODERN -> {
+                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.P)
+                    throw IllegalStateException("Modern cutout only available on Android P or higher")
+                context?.window?.decorView?.rootWindowInsets?.displayCutout?.safeInsetTop ?: 0
+            }
+            CutoutSupport.EXTENDED -> {
+                if (Build.VERSION.SDK_INT < Build.VERSION_CODES.Q)
+                    throw IllegalStateException("Extended cutout only available on Android Q or higher")
+                context?.window?.decorView?.rootWindowInsets?.displayCutout?.boundingRectTop?.height()?.toFloat() ?: 0f
+            }
+            CutoutSupport.LEGACY -> {
+                if (isVivo) {
+                    val insetCompat = context?.window?.decorView?.rootWindowInsets?.let {
+                        WindowInsetsCompat.toWindowInsetsCompat(it)
+                    }
+                    val statusBarHeight = insetCompat?.getInsets(WindowInsetsCompat.Type.statusBars())?.top
+                        ?: 24.dpToPx  // 24dp is "standard" height for Android since Marshmallow
+                    var notchHeight = 32.dpToPx
+                    if (notchHeight < statusBarHeight) {
+                        notchHeight = statusBarHeight
+                    }
+                    notchHeight
+                } else if (isMiui) {
+                    val resourceId = context?.resources?.getIdentifier("notch_height",
+                        "dimen", "android") ?: 0
+                    if (resourceId > 0) {
+                        context?.resources?.getDimensionPixelSize(resourceId) ?: 0
+                    } else {
+                        0
+                    }
+                } else {
+                    0
+                }
+            }
+            else -> 0
+        }
+    }
+
+    enum class CutoutSupport {
+        NONE,
+        LEGACY,  // Pre-Android P, the start of this hell
+        MODERN,  // Android P
+        EXTENDED,  // Android Q
     }
 
     enum class LegacyCutoutMode {