+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 
 ## Other
 -->
+## Additions
+- Added cutout support for some pre-Android P devices
+
+## Fixes
+- Fixed cutout behaviour for Android P